2009-07-10 14:19 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/include/hbgtcore.h   * harbour/source/rtl/hbgtcore.c     * changed BYTE to USHORT in few functions which expected character       for GTs with UNICODE screen buffer
  * harbour/source/rtl/gtcgi/gtcgi.c
  * harbour/source/rtl/gtstd/gtstd.c
  * harbour/source/rtl/gttrm/gttrm.c
    * updated for USHORT used as character holder

  * harbour/source/rtl/gtwvt/gtwvt.c
    % minor modification

  * harbour/source/rtl/gtxwc/gtxwc.c
  * harbour/contrib/gtwvg/gtwvg.c
  * harbour/contrib/gtwvg/wvgcore.c
  * harbour/contrib/gtalleg/gtalleg.c
    * updated for BYTE to int as color holder modification

  * harbour/contrib/hbct/screen1.c
  * harbour/contrib/hbct/screen2.c
  * harbour/contrib/hbct/ctwfunc.c
  * harbour/contrib/hbnf/ftattr.c
    ! fixed casting changing values or added casting to force strict
      translation
    * modified to accept color and character values with wider ranges

  * harbour/contrib/hbct/ctwin.c
    ! removed unnecessary stripping to BYTE in character and color values

  * harbour/source/rtl/console.c
    * pacified warning

+++ harbour/source/rtl/console.c
@@ -357,7 +357,7 @@ HB_FUNC( QOUT )
 
       if( pPrnPos->col )
       {
-         if( pPrnPos->col > sizeof( buf ) )
+         if( pPrnPos->col > ( int ) sizeof( buf ) )
          {
             char * pBuf = ( char * ) hb_xgrab( pPrnPos->col );
             memset( pBuf, ' ', pPrnPos->col );
